feat: 검증 된 Participant 모델 기능 구현
