+++ README.md
@@ -47,7 +47,7 @@
 - 플레이어들
   - [ ] 플레이어들은 플레이어 목록을 가지고 있다.
 - 참여자
-  - [ ] 참여자는 딜러 혹은 플레이어가 된다.
+  - [x] 참여자는 딜러 혹은 플레이어가 된다.
 - 참여자들
   - [ ] 참여자들은 참여자 목록을 가지고 있다.
 ---
